Changes Presentation project to use Avalonia
Changed the presentation project to use Avalonia instead of blazor, and
made small changes to the .gitignore to support Rider development.
