+++ .gitignore
@@ -2,3 +2,4 @@
 */bin
 */obj
 .vscode
+.idea
