+++ NinetyNine.sln
@@ -18,7 +18,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Repository", "Repository\Repository.csproj", "{13734F15-653C-4E6B-A0BD-7C71A20A9E5F}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Presentation", "Presentation\Presentation.csproj", "{1D6F92A2-6CA7-443A-AA82-1614948B0200}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Presentation", "Presentation\Presentation.csproj", "{B8A78904-AD6B-41BA-B50A-35E256CFD738}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -34,10 +34,10 @@ Global
 		{13734F15-653C-4E6B-A0BD-7C71A20A9E5F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{13734F15-653C-4E6B-A0BD-7C71A20A9E5F}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{13734F15-653C-4E6B-A0BD-7C71A20A9E5F}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1D6F92A2-6CA7-443A-AA82-1614948B0200}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1D6F92A2-6CA7-443A-AA82-1614948B0200}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1D6F92A2-6CA7-443A-AA82-1614948B0200}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1D6F92A2-6CA7-443A-AA82-1614948B0200}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B8A78904-AD6B-41BA-B50A-35E256CFD738}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B8A78904-AD6B-41BA-B50A-35E256CFD738}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B8A78904-AD6B-41BA-B50A-35E256CFD738}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B8A78904-AD6B-41BA-B50A-35E256CFD738}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
