+++ Presentation/App.axaml
@@ -1,12 +1,12 @@
 <Application xmlns="https://github.com/avaloniaui"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="using:Avalonia.Presentation"
-             x:Class="Avalonia.Presentation.App">
+             xmlns:local="using:NinetyNine.Presentation"
+             x:Class="NinetyNine.Presentation.App">
     <Application.DataTemplates>
         <local:ViewLocator/>
     </Application.DataTemplates>
 
     <Application.Styles>
-        <FluentTheme Mode="Light"/>
+        <FluentTheme Mode="Dark"/>
     </Application.Styles>
 </Application>
