+++ Presentation/Assets/avalonia-logo.ico
