--- Presentation/Pages/Index.razor
@@ -1,7 +1,0 @@
-﻿@page "/"
-
-<h1>Ninety-Nine</h1>
-<a href="https://github.com/CareyCilyok/NinetyNine/blob/master/README.md">See the description for 99 here.</a>
-<div>
-   <iframe src="./Data/README.md"/>
-</div>
