--- Presentation/Pages/_Host.cshtml
@@ -1,37 +1,0 @@
-﻿@page "/"
-@namespace NinetyNine.Presentation.Pages
-@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
-@{
-    Layout = null;
-}
-
-<!DOCTYPE html>
-<html lang="en">
-<head>
-   <meta charset="utf-8" />
-   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-   <title>99</title>
-   <base href="~/" />
-   <link rel="stylesheet" href="css/bootstrap/bootstrap.min.css" />
-   <link href="css/ScoreSheet.css" rel="stylesheet" />
-   <link href="css/site.css" rel="stylesheet" />
-</head>
-<body>
-    <app>
-        <component type="typeof(App)" render-mode="ServerPrerendered" />
-    </app>
-
-    <div id="blazor-error-ui">
-        <environment include="Staging,Production">
-            An error has occurred. This application may no longer respond until reloaded.
-        </environment>
-        <environment include="Development">
-            An unhandled exception has occurred. See browser dev tools for details.
-        </environment>
-        <a href="" class="reload">Reload</a>
-        <a class="dismiss">🗙</a>
-    </div>
-
-    <script src="_framework/blazor.server.js"></script>
-</body>
-</html>
