--- Presentation/Properties/launchSettings.json
@@ -1,27 +1,0 @@
-{
-  "iisSettings": {
-    "windowsAuthentication": false,
-    "anonymousAuthentication": true,
-    "iisExpress": {
-      "applicationUrl": "http://localhost:51548",
-      "sslPort": 44385
-    }
-  },
-  "profiles": {
-   "IIS Express": {
-      "commandName": "IISExpress",
-      "launchBrowser": true,
-      "environmentVariables": {
-         "ASPNETCORE_ENVIRONMENT": "Development"
-      }
-   },
-   "NinetyNine.Presentation": {
-      "commandName": "Project",
-      "launchBrowser": true,
-      "applicationUrl": "https://localhost:5001;http://localhost:5000",
-      "environmentVariables": {
-         "ASPNETCORE_ENVIRONMENT": "Development"
-      }
-   }
-  }
-}
