--- Presentation/_Imports.razor
@@ -1,9 +1,0 @@
-﻿@using System.Net.Http
-@using Microsoft.AspNetCore.Authorization
-@using Microsoft.AspNetCore.Components.Authorization
-@using Microsoft.AspNetCore.Components.Forms
-@using Microsoft.AspNetCore.Components.Routing
-@using Microsoft.AspNetCore.Components.Web
-@using Microsoft.JSInterop
-@using NinetyNine.Presentation
-@using NinetyNine.Presentation.Shared
