--- Presentation/wwwroot/css/ScoreSheet.css
@@ -1,52 +1,0 @@
-﻿.Inning {
-   display: table;
-}
-
-.Title {
-   display: table-caption;
-   text-align: center;
-   font-weight: bold;
-   font-size: larger;
-}
-
-.InningHeading {
-   display: table-row;
-   border: solid;
-   border-width: thin;
-   font-weight: bold;
-   text-align: center;
-}
-
-.Row {
-   display: table-row;
-}
-
-.InningBreakBonusCell {
-   display: table-cell;
-   border: solid;
-   border-width: thin;
-   text-align: center;
-   padding-left: 5px;
-   padding-right: 5px;
-   border-radius: 4px;
-}
-
-.InningBallCountCell {
-   display: table-cell;
-   border: solid;
-   border-width: thin;
-   text-align: center;
-   padding-left: 5px;
-   padding-right: 5px;
-}
-
-.RunningBallCountCell {
-   display: table-cell;
-   border: solid;
-   border-width: thin;
-   text-align: center;
-   font-weight: bold;
-   padding-left: 5px;
-   padding-right: 5px;
-}
-
